Split the tidy_fetuses file into 5 separate files based on the sections.
I think it is more manageable for the students to deal with five smaller exercises than one large. This gives us an opportunity to repeat some of the important stuff. So I have split the large file into five smaller, and repeated the most importent from the previous files in the beginning of each, so the code runs (File 1-3). File 4 and 5 are exercises and solutions.